requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.7
Flask-CLI==0.4.0
python-dotenv==1.0.1
PyMuPDF==1.24.9
mysqlclient==2.2.4
google-cloud-vision==3.7.4